requirements.txt
# Selenium WebDriver
selenium==4.15.2
webdriver-manager==4.0.1

# Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
pytest-mock==3.12.0

# Allure Reporting
allure-pytest==2.13.2

# Data Handling
openpyxl==3.1.2
pandas==2.1.3
faker==20.1.0

# Configuration and Logging
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1

# API Testing (optional)
requests==2.31.0

# Utilities
pillow==10.1.0
